prepare for coupons trigger
